+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Align augmented generation with camelCase (Thanks to https://github.com/monitz87)
 - Fix bitfield endian encoding (Thanks to investigation by https://github.com/xlc)
 - Fix entries decoding for `Option<Number>` (& simplify, decoding correctly on rpc layer)
+- Add `rpc.system.{localPeerId, localListenAddresses}`
 - Ensure proper propagation of `api.runtimeVersion` (used for signing) after upgrade
 - Field ordering fix for deprecated `Linkage` (used in linked maps)
 - `derive.account.info` now returns all field types (not only raw)